+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Tab Override jQuery Plugin Changelog
 
+## 4.0.0-dev
+* Use `handlers` namespace
+* Added `addDelegatedExtension()` method
+
 ## 3.2.4 (2013-03-12)
 * Added CommonJS module support
 * Updated build process